Move _stopped() into a trait in preparation for sharing with Edit
